app.py
"""Streamlit chatbot front‑end for BreedSpotter."""
from __future__ import annotations

import io
from pathlib import Path

import streamlit as st
from PIL import Image

from breedspotter.classifier import load_classifier
from breedspotter.data import load_metadata, load_breed_profiles
from breedspotter.llm import describe_breed

st.set_page_config(page_title="BreedSpotter 🐶", page_icon="🐾")

st.title("🐾 BreedSpotter")
st.markdown("Prześlij zdjęcie psa, a powiem Ci, jaka to rasa — i opowiem o niej w kilku zdaniach.")

# Lazy init
if "_init" not in st.session_state:
    st.session_state._init = True
    _df, st.session_state.breeds = load_metadata()
    st.session_state.profiles = load_breed_profiles()
    st.session_state.clf = load_classifier(st.session_state.breeds)

uploaded = st.file_uploader("Wybierz zdjęcie", type=["jpg", "jpeg", "png"])
if uploaded:
    with st.spinner("Analizuję obraz…"):
        img = Image.open(io.BytesIO(uploaded.read())).convert("RGB")
        breed, prob, ranked = st.session_state.clf.predict(img)
    st.image(img, caption=f"Najbardziej prawdopodobna rasa: **{breed}** ({prob*100:.1f}%)", use_container_width=True)

    # Generate / fetch description
    profile = st.session_state.profiles.get(breed, "Brak opisu w bazie.")
    description = describe_breed(breed, profile)

    st.markdown(f"### Opis rasy: {breed}")
    st.write(description)

    if st.toggle("Pokaż 5 najlepszych typów"):
        for b, p in sorted(ranked, key=lambda t: t[1], reverse=True)[:5]:
            st.write(f"• {b}: {p*100:.1f}%")