requirements.txt
streamlit==1.34.0
pillow>=10.3.0
toml>=0.10.2
python-dotenv>=1.0.1

numpy>=1.19.3
pandas>=2.2.2
scikit-learn>=1.5.0
networkx>=3.3.1

torch==2.3.0
torchvision==0.18.0
timm>=0.9.16

transformers==4.41.2
accelerate>=0.27.2
peft==0.10.0

openai>=1.23.0
langchain-core==0.3.66
langchain-openai==0.3.24
langchain==0.3.26

fastapi>=0.110.0
pydantic>=2.11.7